Export user_monsters queries and import into database file.
